Initial structure of env.py with Gym wrapper
This work was done together over Discord

We decided to structure the Gymnasium env to interface with the
game's GameState and GameStateData objects. Current goal is to
get a functioning Pac-Man agent that can play against an A* ghost,
with room left to select different agents in the future.

Hamza had a great idea to use a 'Master Agent' to control all ghosts
at once, which would hopefully allow for cooperative strategies
amongst those agents.

Kaya also spent time writing out our initial objective function scoring
and the weights attributed to a given action - implementation of that
for the step() function is our next objective.